Time: 92 ms (98.33%) | Memory: 108.3 MB (96.38%) - LeetSync

+++ leetcode/1631-number-of-sub-arrays-with-odd-sum/number-of-sub-arrays-with-odd-sum.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int numOfSubarrays(vector<int>& arr) {
-        long int MOD = 1e9 + 7;
+        long long MOD = 1e9 + 7;
         long long ans = 0, count = 0, even = 0, odd = 0;
 
         for (int x : arr) {
